refactor : atoms
