+++ src/components/atoms/BackBtn/BackBtn.jsx
@@ -1,6 +1,15 @@
 import { useNavigate } from "react-router-dom";
 import { ReactComponent as BackIcon } from "assets/receiptPage/back_icon.svg";
-import * as S from "./style";
+import * as S from "./BackBtn.styles";
+
+/**
+ * BackBtn
+ *
+ * 뒤로가기 버튼입니다
+ * 누르면 이전 페이지로 이동합니다
+ *
+ * @returns
+ */
 export function BackBtn() {
   const navigate = useNavigate();
   const onClickBack = () => {
