+++ src/components/atoms/BackBtn/index.js
@@ -1,0 +1,1 @@
+export { BackBtn } from "./BackBtn";
