+++ src/components/atoms/DayOfWeek/index.js
@@ -1,0 +1,1 @@
+export { DayOfWeek } from "./DayOfWeek";
