+++ src/components/atoms/HeaderText/HeaderText.jsx
@@ -1,4 +1,4 @@
-import * as S from "./style";
+import * as S from "./HeaderText.styles";
 
 /**
  * Header component
