+++ src/components/atoms/HeaderText/index.js
@@ -1,0 +1,1 @@
+export { HeaderText } from "./HeaderText";
