+++ src/components/atoms/Portal/index.js
@@ -1,0 +1,1 @@
+export { Portal } from "./Portal";
