+++ src/components/atoms/ReceiptBarcode/ReceiptBarcode.styles.jsx
@@ -1,0 +1,18 @@
+import styled from "styled-components";
+
+export const BarcodeContainer = styled.div`
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  margin: 24px 0;
+  transform: translateY(8px);
+`;
+
+export const Barcode = styled.div`
+  opacity: 84%;
+  transform: scale(110%);
+`;
+
+export const Address = styled.div`
+  margin-top: 8px;
+`;
