+++ src/components/atoms/ReceiptPaperInfo/ReceiptPaperInfo.styles.jsx
@@ -1,0 +1,6 @@
+import styled from "styled-components";
+
+export const Info = styled.div`
+  transform: translateX(4px);
+  align-self: flex-start;
+`;
