+++ src/components/atoms/ReceiptPaperTriangle/index.js
@@ -1,0 +1,1 @@
+export { ReceiptPaperTriangle } from "./ReceiptPaperTriangle";
