+++ src/components/atoms/ReceiptQuotes/ReceiptQuotes.styles.jsx
@@ -1,0 +1,7 @@
+import styled from "styled-components";
+
+export const Quotes = styled.div`
+  width: 92%;
+  margin: 24px 0;
+  text-align: center;
+`;
