+++ src/components/atoms/ReceiptQuotes/index.js
@@ -1,0 +1,1 @@
+export { ReceiptQuotes } from "./ReceiptQuotes";
