+++ src/components/atoms/StartTimerBtn/index.js
@@ -1,0 +1,1 @@
+export { StartTimerBtn } from "./StartTimerBtn";
