--- src/components/atoms/Timer/index.jsx
@@ -1,3 +1,0 @@
-/**
- * 만들어놓긴 했는데 어떻게 활용해야 할지 ㅎㅎ 감이 잘 안잡히네여..b
- */
