+++ src/components/atoms/TimerImage/TimerImage.styles.jsx
