+++ src/components/atoms/TimerImage/index.js
@@ -1,0 +1,1 @@
+export { TimerImage } from "./TimerImage";
