+++ src/components/molecules/ReceiptPaper/index.js
@@ -1,0 +1,1 @@
+export { ReceiptPaper } from "./ReceiptPaper";
