+++ src/components/molecules/ReceiptTodo/ReceiptTodo.styles.jsx
@@ -1,0 +1,19 @@
+import styled from "styled-components";
+
+export const Container = styled.div`
+  width: 100%;
+`;
+
+export const Info = styled.div``;
+export const Todo = styled.div`
+  display: flex;
+  justify-content: space-between;
+`;
+export const Task = styled.div`
+  width: 100%;
+  white-space: nowrap;
+  overflow: hidden;
+  text-overflow: ellipsis;
+`;
+
+export const Timer = styled.div``;
