+++ src/components/molecules/ReceiptTodo/index.js
@@ -1,0 +1,1 @@
+export { ReceiptTodo } from "./ReceiptTodo";
